+++ src/ui.rs
@@ -9,6 +9,7 @@ mod dialog_about;
 mod dialog_archive_prefix;
 mod dialog_device_missing;
 mod dialog_encryption_password;
+mod dialog_exclude;
 mod dialog_info;
 mod dialog_prune;
 mod dialog_prune_review;