+++ src/ui/builder.rs
@@ -657,6 +657,72 @@ impl DialogEncryptionPassword {
 }
 
 #[derive(Clone)]
+pub struct DialogExclude {
+    builder: gtk::Builder,
+}
+
+#[derive(Clone)]
+pub struct DialogExcludeWeak {
+    builder: glib::WeakRef<gtk::Builder>,
+}
+
+impl glib::clone::Downgrade for DialogExclude {
+    type Weak = DialogExcludeWeak;
+
+    fn downgrade(&self) -> Self::Weak {
+        Self::Weak {
+            builder: self.builder.downgrade(),
+        }
+    }
+}
+
+impl glib::clone::Upgrade for DialogExcludeWeak {
+    type Strong = DialogExclude;
+
+    fn upgrade(&self) -> Option<Self::Strong> {
+        Some(Self::Strong {
+            builder: self.builder.upgrade()?,
+        })
+    }
+}
+
+impl DialogExclude {
+    pub fn new() -> Self {
+        Self {
+            builder: gtk::Builder::from_string(include_str!(concat!(
+                env!("CARGO_MANIFEST_DIR"),
+                "/src/ui/dialog_exclude.ui"
+            ))),
+        }
+    }
+
+    fn get<T: glib::IsA<glib::object::Object>>(&self, id: &str) -> T {
+        gtk::Builder::object(&self.builder, id).unwrap_or_else(|| {
+            panic!(
+                "Object with id '{}' not found in 'src/ui/dialog_exclude.ui'",
+                id
+            )
+        })
+    }
+
+    pub fn dialog(&self) -> adw::Window {
+        self.get("dialog")
+    }
+
+    pub fn exclude_file(&self) -> adw::ActionRow {
+        self.get("exclude_file")
+    }
+
+    pub fn exclude_folder(&self) -> adw::ActionRow {
+        self.get("exclude_folder")
+    }
+
+    pub fn suggestions(&self) -> adw::PreferencesGroup {
+        self.get("suggestions")
+    }
+}
+
+#[derive(Clone)]
 pub struct DialogPrune {
     builder: gtk::Builder,
 }