+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Change to hide advanced settings behind toggle button in setup
 - Change to use AdwAboutWindow
 - Change to use AdwMessageDialog
+- Change to symbolic icons in include/exclude list
+- Add predefined exclusion rules
 
 ## 0.4.2 (2022-07-12)